requirements.txt
numpy~=1.26.4
pillow~=10.4.0
torchvision~=0.18.1
distinctipy~=1.3.4
matplotlib~=3.9.1
pandas~=2.2.2